i18n-devstack: check if /opt/stack/horizon exists before cleanup

+++ roles/i18n-devstack/templates/99-horizon-i18n.sh
@@ -4,10 +4,12 @@ if is_service_enabled horizon; then
     if [[ "$1" == "source" ]]; then
         :
     elif [[ "$1" == "stack" && "$2" == "pre-install" ]]; then
-        cd $DEST/horizon
-        rm -rf horizon/locale
-        rm -rf openstack_dashboard/locale
-        git checkout -- .
+        if [ -d $DEST/horizon ]; then
+            cd $DEST/horizon
+            rm -rf horizon/locale
+            rm -rf openstack_dashboard/locale
+            git checkout -- .
+        fi
     elif [[ "$1" == "stack" && "$2" == "install" ]]; then
         :
     elif [[ "$1" == "stack" && "$2" == "post-config" ]]; then
